Fix goal problem
Hi, I made everything work, I just need to know if you have a detail that happens when you do an update, and you do not touch the dates, it is like a day is subtracted.
Also in the backend when you do a get, the comments and planned finish data are missing, please make sure to send all the data correctly.